+++ src/elements/index.tsx
@@ -1,0 +1,6 @@
+export { default as Cover } from './Cover';
+export { default as Grid } from './Grid';
+export { default as GridItem } from './GridItem';
+export { default as GridHeader } from './GridHeader';
+export { default as Main } from './Main';
+export { default as Section } from './Section';
